+++ packages/openactive-integration-tests/test/features/README.md
@@ -72,9 +72,9 @@ The tests for these features are fully stubbed, and are not yet implemented.
 | payment | Payment reconciliation detail validation ([payment-reconciliation-detail-validation](./payment/payment-reconciliation-detail-validation/README.md)) | [Optional](https://www.openactive.io/open-booking-api/EditorsDraft/#payment-reconciliation-detail-validation) | Booking with valid, invalid, and missing Payment details | [TestOpportunityBookableUsingPayment](https://openactive.io/test-interface#TestOpportunityBookableUsingPayment) x15, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x5 |
 | restrictions | Booking restrictions ([booking-restrictions](./restrictions/booking-restrictions/README.md)) | [Optional](https://www.openactive.io/open-booking-api/EditorsDraft/#booking-restrictions) | Support for genderRestriction, ageRestriction and additionalAdmissionRestriction |  |
 | restrictions | validFromBeforeStartDate booking window ([booking-window](./restrictions/booking-window/README.md)) | [Optional](https://www.openactive.io/open-booking-api/EditorsDraft/#definition-of-a-bookable-opportunity-and-offer-pair) | Duration of window before an opportunity where it is bookable | [TestOpportunityBookableWithinValidFromBeforeStartDate](https://openactive.io/test-interface#TestOpportunityBookableWithinValidFromBeforeStartDate) x3, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x2, [TestOpportunityBookableOutsideValidFromBeforeStartDate](https://openactive.io/test-interface#TestOpportunityBookableOutsideValidFromBeforeStartDate) x3 |
-| tax | Business-to-business Tax Calculation (TaxGross) ([business-to-business-tax-calculation-gross](./tax/business-to-business-tax-calculation-gross/README.md)) | [Optional](https://www.openactive.io/open-booking-api/EditorsDraft/#business-to-business-tax-calculation-by-booking-system-is-optional) | Tax calculation when the customer is of type Organization (business-to-business), when the seller has taxMode TaxGross | [TestOpportunityBookablePaidTaxGross](https://openactive.io/test-interface#TestOpportunityBookablePaidTaxGross) x3, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x1 |
-| tax | Business-to-business Tax Calculation (TaxNet) ([business-to-business-tax-calculation-net](./tax/business-to-business-tax-calculation-net/README.md)) | [Optional](https://www.openactive.io/open-booking-api/EditorsDraft/#business-to-business-tax-calculation-by-booking-system-is-optional) | Tax calculation when the customer is of type Organization (business-to-business), when the seller has taxMode TaxNet | [TestOpportunityBookablePaidTaxNet](https://openactive.io/test-interface#TestOpportunityBookablePaidTaxNet) x3, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x1 |
-| tax | Business-to-consumer Tax Calculation (TaxGross) ([business-to-consumer-tax-calculation-gross](./tax/business-to-consumer-tax-calculation-gross/README.md)) | [Optional](https://www.openactive.io/open-booking-api/EditorsDraft/#business-to-consumer-tax-calculation-by-booking-system-is-mandatory) | Tax calculation when the customer is of type Person (business-to-consumer), when the seller has taxMode TaxGross | [TestOpportunityBookablePaidTaxGross](https://openactive.io/test-interface#TestOpportunityBookablePaidTaxGross) x3, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x1 |
+| tax | Business-to-business Tax Calculation (TaxGross) ([business-to-business-tax-calculation-gross](./tax/business-to-business-tax-calculation-gross/README.md)) | [Optional](https://www.openactive.io/open-booking-api/EditorsDraft/#business-to-business-tax-calculation-by-booking-system-is-optional) | Tax calculation when the customer is of type Organization (business-to-business), when the seller has taxMode TaxGross | [TestOpportunityBookableNonFreeTaxGross](https://openactive.io/test-interface#TestOpportunityBookableNonFreeTaxGross) x3, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x1 |
+| tax | Business-to-business Tax Calculation (TaxNet) ([business-to-business-tax-calculation-net](./tax/business-to-business-tax-calculation-net/README.md)) | [Optional](https://www.openactive.io/open-booking-api/EditorsDraft/#business-to-business-tax-calculation-by-booking-system-is-optional) | Tax calculation when the customer is of type Organization (business-to-business), when the seller has taxMode TaxNet | [TestOpportunityBookableNonFreeTaxNet](https://openactive.io/test-interface#TestOpportunityBookableNonFreeTaxNet) x3, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x1 |
+| tax | Business-to-consumer Tax Calculation (TaxGross) ([business-to-consumer-tax-calculation-gross](./tax/business-to-consumer-tax-calculation-gross/README.md)) | [Optional](https://www.openactive.io/open-booking-api/EditorsDraft/#business-to-consumer-tax-calculation-by-booking-system-is-mandatory) | Tax calculation when the customer is of type Person (business-to-consumer), when the seller has taxMode TaxGross | [TestOpportunityBookableNonFreeTaxGross](https://openactive.io/test-interface#TestOpportunityBookableNonFreeTaxGross) x3, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x1 |
 | tax | Business-to-consumer Tax Calculation (TaxNet) ([business-to-consumer-tax-calculation-net](./tax/business-to-consumer-tax-calculation-net/README.md)) | [Optional](https://www.openactive.io/open-booking-api/EditorsDraft/#business-to-consumer-tax-calculation-by-booking-system-is-mandatory) | Tax calculation when the customer is of type Person (business-to-consumer), when the seller has taxMode TaxNet |  |
 | terms | termsOfService without requiresExplicitConsent ([terms-of-service](./terms/terms-of-service/README.md)) | [Optional](https://www.openactive.io/open-booking-api/EditorsDraft/#delivery-of-terms-and-conditions-and-privacy-policy) | Displaying terms of service, without requiring consent |  |
 | terms | termsOfService with requiresExplicitConsent ([terms-of-service-with-consent](./terms/terms-of-service-with-consent/README.md)) | [Optional](https://www.openactive.io/open-booking-api/EditorsDraft/#delivery-of-terms-and-conditions-and-privacy-policy) | Displaying terms of service that require consent |  |
