+++ packages/openactive-integration-tests/test/features/tax/business-to-business-tax-calculation-gross/implemented/business-to-business-tax-calculation-gross-test.js
@@ -8,7 +8,7 @@ FeatureHelper.describeFeature(module, {
   testIdentifier: 'business-to-business-tax-calculation-gross',
   testName: 'Tax calculations',
   testDescription: 'The totals of totalPaymentTax should match the sum of the unitTaxSpecification, and the totalPaymentDue.price should equal the total of all acceptedOffer.price PLUS TotalPaymentTax.price.',
-  testOpportunityCriteria: 'TestOpportunityBookablePaidTaxGross',
+  testOpportunityCriteria: 'TestOpportunityBookableNonFreeTaxGross',
   controlOpportunityCriteria: 'TestOpportunityBookable',
 },
 grossTest({ c2ReqTemplateRef: 'businessCustomer', bReqTemplateRef: 'businessCustomer' }));
