+++ packages/openactive-integration-tests/test/features/tax/business-to-business-tax-calculation-net/README.md
@@ -10,7 +10,7 @@ Coverage Status: **none**
 ### Test prerequisites
 Opportunities that match the following criteria must exist in the booking system (for each configured `bookableOpportunityTypesInScope`) for the configured primary Seller in order to use `useRandomOpportunities: true`. Alternatively the following `testOpportunityCriteria` values must be supported by the [test interface](https://openactive.io/test-interface/) of the booking system for `useRandomOpportunities: false`.
 
-[TestOpportunityBookablePaidTaxNet](https://openactive.io/test-interface#TestOpportunityBookablePaidTaxNet) x3, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x1
+[TestOpportunityBookableNonFreeTaxNet](https://openactive.io/test-interface#TestOpportunityBookableNonFreeTaxNet) x3, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x1
 
 *Note the test coverage for this feature is currently nonexistent. The test suite does not yet include non-stubbed tests for this feature.*
 
@@ -29,6 +29,6 @@ Update `default.json` within `packages/openactive-integration-tests/config/` as 
 
 | Identifier | Name | Description | Prerequisites per Opportunity Type |
 |------------|------|-------------|---------------|
-| [business-to-business-tax-calculation-net](./implemented/business-to-business-tax-calculation-net-test.js) | Tax calculations | The totals of totalPaymentTax should match the sum of the unitTaxSpecification, and the totalPaymentDue.price should equal the total of all acceptedOffer.price PLUS TotalPaymentTax.price. | [TestOpportunityBookablePaidTaxNet](https://openactive.io/test-interface#TestOpportunityBookablePaidTaxNet) x3, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x1 |
+| [business-to-business-tax-calculation-net](./implemented/business-to-business-tax-calculation-net-test.js) | Tax calculations | The totals of totalPaymentTax should match the sum of the unitTaxSpecification, and the totalPaymentDue.price should equal the total of all acceptedOffer.price PLUS TotalPaymentTax.price. | [TestOpportunityBookableNonFreeTaxNet](https://openactive.io/test-interface#TestOpportunityBookableNonFreeTaxNet) x3, [TestOpportunityBookable](https://openactive.io/test-interface#TestOpportunityBookable) x1 |
