+++ packages/test-interface-criteria/built-types/criteria/TestOpportunityBookableNonFreeTaxGross.d.ts
@@ -1,0 +1,5 @@
+export type OpportunityConstraint = (opportunity: import("../types/Opportunity").Opportunity, options?: import("../types/Options").Options) => boolean;
+/**
+ * Implements https://openactive.io/test-interface#TestOpportunityBookableNonFreeTaxGross.
+ */
+export const TestOpportunityBookableNonFreeTaxGross: import("../types/Criteria").Criteria;
