+++ packages/test-interface-criteria/src/criteria/TestOpportunityBookableNonFreeTaxNet.js
@@ -8,26 +8,26 @@ const { TestOpportunityBookableNonFree } = require('./TestOpportunityBookableNon
 /**
  * @type {OpportunityConstraint}
  */
-function sellerTaxModeGross(opportunity) {
+function sellerTaxModeNet(opportunity) {
   switch (opportunity['@type']) {
     case 'ScheduledSession':
-      return opportunity.superEvent.organizer.taxMode === 'https://openactive.io/TaxGross';
+      return opportunity.superEvent.organizer.taxMode === 'https://openactive.io/TaxNet';
     case 'Slot':
-      return opportunity.facilityUse.provider.taxMode === 'https://openactive.io/TaxGross';
+      return opportunity.facilityUse.provider.taxMode === 'https://openactive.io/TaxNet';
     default:
       throw new Error(`Type ${opportunity['@type']} not supported`);
   }
 }
 
 /**
- * Implements https://openactive.io/test-interface#TestOpportunityBookablePaidTaxGross.
+ * Implements https://openactive.io/test-interface#TestOpportunityBookableNonFreeTaxNet.
  */
-const TestOpportunityBookablePaidTaxGross = createCriteria({
-  name: 'TestOpportunityBookablePaidTaxGross',
+const TestOpportunityBookableNonFreeTaxNet = createCriteria({
+  name: 'TestOpportunityBookableNonFreeTaxNet',
   opportunityConstraints: [
     [
-      'Seller Tax Mode Gross',
-      sellerTaxModeGross,
+      'Seller Tax Mode Net',
+      sellerTaxModeNet,
     ],
   ],
   offerConstraints: [],
@@ -35,5 +35,5 @@ const TestOpportunityBookablePaidTaxGross = createCriteria({
 });
 
 module.exports = {
-  TestOpportunityBookablePaidTaxGross,
+  TestOpportunityBookableNonFreeTaxNet,
 };
