+++ packages/test-interface-criteria/src/criteria/index.js
@@ -9,8 +9,8 @@ const { TestOpportunityBookableCancellable } = require('./TestOpportunityBookabl
 const { TestOpportunityNotBookableViaAvailableChannel } = require('./TestOpportunityNotBookableViaAvailableChannel');
 const { TestOpportunityBookableOutsideValidFromBeforeStartDate } = require('./TestOpportunityBookableOutsideValidFromBeforeStartDate');
 const { TestOpportunityBookableFlowRequirementOnlyApproval } = require('./TestOpportunityBookableFlowRequirementOnlyApproval');
-const { TestOpportunityBookablePaidTaxNet } = require('./TestOpportunityBookablePaidTaxNet');
-const { TestOpportunityBookablePaidTaxGross } = require('./TestOpportunityBookablePaidTaxGross');
+const { TestOpportunityBookableNonFreeTaxNet } = require('./TestOpportunityBookableNonFreeTaxNet');
+const { TestOpportunityBookableNonFreeTaxGross } = require('./TestOpportunityBookableNonFreeTaxGross');
 const { TestOpportunityBookableNonFreePrepaymentOptional } = require('./TestOpportunityBookableNonFreePrepaymentOptional');
 const { TestOpportunityBookableNonFreePrepaymentUnavailable } = require('./TestOpportunityBookableNonFreePrepaymentUnavailable');
 const { TestOpportunityBookableNonFreePrepaymentRequired } = require('./TestOpportunityBookableNonFreePrepaymentRequired');
@@ -30,8 +30,8 @@ module.exports = {
     TestOpportunityNotBookableViaAvailableChannel,
     TestOpportunityBookableOutsideValidFromBeforeStartDate,
     TestOpportunityBookableFlowRequirementOnlyApproval,
-    TestOpportunityBookablePaidTaxNet,
-    TestOpportunityBookablePaidTaxGross,
+    TestOpportunityBookableNonFreeTaxNet,
+    TestOpportunityBookableNonFreeTaxGross,
     TestOpportunityBookableNonFreePrepaymentOptional,
     TestOpportunityBookableNonFreePrepaymentUnavailable,
     TestOpportunityBookableNonFreePrepaymentRequired,
